Restructure copy module
