+++ iomete-lakehouse-backup/src/main/kotlin/com/iomete/backup/copy/internal/PathResolver.kt
@@ -1,23 +1,6 @@
-package com.iomete.backup.copy
-
-import com.iomete.backup.config.S3Config
-import com.iomete.backup.config.StorageConfig
+package com.iomete.backup.copy.internal
 
 object PathResolver {
-    fun resolveRootUri(config: StorageConfig): String =
-        when (config) {
-            is S3Config -> resolveS3Root(config)
-        }
-
-    private fun resolveS3Root(config: S3Config): String {
-        val prefix = config.prefix.trim('/')
-        return if (prefix.isEmpty()) {
-            "s3a://${config.bucket}"
-        } else {
-            "s3a://${config.bucket}/$prefix"
-        }
-    }
-
     fun resolveTargetPath(
         sourceFilePath: String,
         sourceRoot: String,
